Moved home page to a separate dir. Added test page. #1
