+++ src/pages/home/HomePage.module.scss
@@ -1,4 +1,4 @@
-@import "../sass/colors";
+@import "../../sass/colors";
 
 .root {
     min-height: 100vh;
