+++ src/pages/index.ts
@@ -1,3 +1,3 @@
-import {default as HomePage} from "./HomePage";
+import {default as HomePage} from "./home/HomePage";
 
 export default HomePage;
